fake_vagrant_sshd fixes.
The big one is handling '' cmds immediately, instead of waiting
for the client to say something.  This makes boot up way faster, and
reliable.

A smaller one is to flush stdout on print.  Apparently python doesn't
switch to unbuffered when you redirect to a file and so nothing shows
up in the logfile until way later.  Not sure what is going on there.